Move codes to view model
